@@ -6,6 +6,17 @@
     <artifactId>spring-root</artifactId>
     <version>1.0</version>
     <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.11.0</version>
+                <configuration>
+                    <source>21</source>
+                    <target>21</target>
+                </configuration>
+            </plugin>
+        </plugins>
     </build>
     <parent>
         <groupId>org.springframework.boot</groupId>
@@ -16,7 +27,7 @@
         <mysql.version>8.0.33</mysql.version>
         <sitemesh.version>2.4.2</sitemesh.version>
         <displaytag.version>1.2</displaytag.version>
-        <java.version>1.8</java.version>
+        <java.version>21</java.version>
         <jackson.version>2.15.3</jackson.version>
     </properties>
     <dependencies>
