@@ -24,7 +24,7 @@ public class JpaAuditingConfig {
         public Optional<String> getCurrentAuditor() {
             Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
             if (authentication == null || !authentication.isAuthenticated()) {
-                return Optional.empty(); // Return an empty Optional instead of null
+                return Optional.empty();
             }
             return Optional.of(authentication.getName());
         }
