@@ -1,27 +1,20 @@
 package com.javaweb.converter;
 
-import com.javaweb.constant.AppConstant;
 import com.javaweb.entity.MediaEntity;
 import com.javaweb.entity.RoleEntity;
 import com.javaweb.entity.UserEntity;
 import com.javaweb.enums.AccountStatus;
 import com.javaweb.enums.MediaType;
 import com.javaweb.enums.RoleType;
-import com.javaweb.exception.EntityNotFoundException;
-import com.javaweb.model.dto.RoleDTO;
 import com.javaweb.model.dto.UserDTO;
 import com.javaweb.model.request.UserRequestDTO;
-import com.javaweb.repository.MediaRepository;
 import com.javaweb.repository.RoleRepository;
 import com.javaweb.utils.FileUtil;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Component;
 
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.Objects;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Component
@@ -29,16 +22,14 @@ import java.util.stream.Collectors;
 public class UserConverter implements EntityConverter<UserEntity, UserRequestDTO, UserDTO> {
 
     private final ModelMapper modelMapper;
-    private final RoleConverter roleConverter;
     private final RoleRepository roleRepository;
-    private final MediaRepository mediaRepository;
 
     public UserDTO toDTO(UserEntity entity) {
         UserDTO result = modelMapper.map(entity, UserDTO.class);
-        Set<RoleDTO> roleDTOS = entity.getRoles() != null ? entity.getRoles().stream().map(roleConverter::toDTO).collect(Collectors.toCollection(HashSet::new)) : new HashSet<>();
+        Set<String> roleDTOS = entity.getRoles() != null ? entity.getRoles().stream()
+                .map(role -> "ROLE_" + role.getCode())
+                .collect(Collectors.toCollection(LinkedHashSet::new)) : null;
         result.setRoles(roleDTOS);
-        String roleNames = entity.getRoles() != null ? entity.getRoles().stream().map(RoleEntity::getName).collect(Collectors.joining(",")) : "";
-        result.setVisibleRoles(roleNames);
         return result;
     }
 
@@ -48,9 +39,7 @@ public class UserConverter implements EntityConverter<UserEntity, UserRequestDTO
         MediaEntity media = new MediaEntity();
         media.setFileType(MediaType.IMAGE);
         if (userRequestDTO.getAvatar() == null) {
-            media = mediaRepository.findByFileUrl(AppConstant.DEFAULT_USER_AVT_PATH)
-                    .orElseThrow(() -> new EntityNotFoundException("Default avatar not found!"));
-            result.setAvatar(media);
+            result.setAvatar(null);
         }
 
         if (userRequestDTO.getRequestRoles() != null && !userRequestDTO.getRequestRoles().isEmpty()) {
